ksql-parser/src/test/java/io/confluent/ksql/parser/tree/SessionWindowExpressionTest.java
/*
 * Copyright 2018 Confluent Inc.
 *
 * Licensed under the Confluent Community License (the "License"); you may not use
 * this file except in compliance with the License.  You may obtain a copy of the
 * License at
 *
 * http://www.confluent.io/confluent-community-license
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
 * WARRANTIES OF ANY KIND, either express or implied.  See the License for the
 * specific language governing permissions and limitations under the License.
 */

package io.confluent.ksql.parser.tree;

import static org.hamcrest.MatcherAssert.assertThat;
import static org.hamcrest.Matchers.is;

import io.confluent.ksql.execution.windows.SessionWindowExpression;
import io.confluent.ksql.model.WindowType;
import io.confluent.ksql.serde.WindowInfo;
import java.util.Optional;
import java.util.concurrent.TimeUnit;
import org.junit.Before;
import org.junit.Test;
import org.junit.runner.RunWith;
import org.mockito.junit.MockitoJUnitRunner;

@RunWith(MockitoJUnitRunner.class)
public class SessionWindowExpressionTest {

  private SessionWindowExpression windowExpression;

  @Before
  public void setUp() {
    windowExpression = new SessionWindowExpression(5, TimeUnit.SECONDS);
  }

  @Test
  public void shouldReturnWindowInfo() {
    assertThat(windowExpression.getWindowInfo(),
        is(WindowInfo.of(WindowType.SESSION, Optional.empty())));
  }
}